requirements.txt
pdf2image==1.14.0
numpy==1.18.5
Pillow==8.0.1
